% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMITE_calibration.R
\name{SMITE.calib}
\alias{SMITE.calib}
\title{SMITE Calibration}
\usage{
SMITE.calib(
  A,
  b,
  Ae = NULL,
  be = NULL,
  it = 10000,
  noise = "white",
  acc = NULL,
  eigenclean = NULL,
  alpha = 0.05
)
}
\arguments{
\item{A}{The 'forward' coral variable matrix (t x p; time in rows, coral parameters in columns).}

\item{b}{The reconstruction target (t x 1).}

\item{Ae}{A t x p matrix containing the errors for the forward matrix.}

\item{be}{Error estimates for the reconstruction target. It can either be one value or of equal length to the reconstruction target (t x 1).}

\item{it}{The number of bootstrap Monte Carlo iterations (recommended at 10,000).}

\item{noise}{Either 'white' or 'red', describing how the noise is propagated into the model.}

\item{acc}{Autocorrelation Coefficient; if noise is specified as 'red', it describes the degree of autocorrelation in the error term.}

\item{eigenclean}{Describes how (if at all) the singular values should be truncated to 'clean' the inverse solution. If the value is between 0 and 1, it will remove singular values based on the cumulative variance explained. If the value is greater than 1, it will return that many singular values (from the highest order).}

\item{alpha}{The significance level for the confidence interval (i.e., 0.05 = 95-percent confidence).}
}
\value{
S - Confidence intervals on the singular values of the forward matrix,

m - SMITE model parameters, or loadings, for each column of the forward matrix.

bhat - Predicted target anomalies

recon - Predicted absolute target values

e - Relevant error metrics. It enumerates (half) the distance of the 95-percent confidence interval (the average distance from the mean to the upper and lower bounds) for the singular values and the model parameters. It also enumerates the Standard Error of Prediction (SEP), Root-Mean-Square-Error (RMSE), and the correlation coefficient (r).
}
\description{
Create a model (m) based on the 'forward' coral variable matrix (A) and the reconstruction target (b).
}
