% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMITEcalibration.R
\name{SMITE.calib}
\alias{SMITE.calib}
\title{SMITE Calibration}
\usage{
SMITE.calib(
  A,
  b,
  Ae = NULL,
  be = NULL,
  it = 10000,
  eigenclean = NULL,
  alpha = 0.05,
  xval = NULL,
  weights = TRUE,
  itj = 25,
  tol = 1e-08
)
}
\arguments{
\item{A}{The 'forward' coral variable matrix (t x p; time in rows, coral parameters in columns).}

\item{b}{The reconstruction target (t x 1).}

\item{Ae}{A t x p matrix containing the errors for the forward matrix.}

\item{be}{Error estimates for the reconstruction target. It can either be one value or a vector of equal length to the reconstruction target (t x 1).}

\item{it}{The number of bootstrap Monte Carlo iterations (recommended at 10,000).}

\item{eigenclean}{Describes how (if at all) the singular values should be truncated to 'clean' the inverse solution. If the value is between 0 and 1, it will remove singular values based on the cumulative variance explained. If the value is greater than 1, it will return that many singular values (from the highest order).}

\item{alpha}{The significance level for the confidence interval (i.e., 0.05 = 95-percent confidence).}

\item{xval}{Cross-validation window size.}

\item{weights}{TRUE or FALSE on whether you want to incorporate uncertainty (Ae and be) into calibration via an iterative Jacobian reweighting loop. Default is TRUE.}
}
\value{
S - Confidence intervals on the singular values of the forward matrix,

x - Summary of SMITE model parameters, or loadings, for each column of the forward matrix.

x_bootstrap - Individual bootstrap iterations of SMITE model parameters, needed for propagating uncertainty in predictions.

res_model - Standard deviation and autoregressive coefficients (1 & 2) of the residuals of each bootstrap.

bhat - Predicted target values from in-sample blocks

bhat_xval - predicted target values from out-of-sample blocks

e_stats - Confidence intervals for R and RMSE (non-cross-validated and cross-validated). The size of the confidence interval is based on alpha.
}
\description{
Create a model (x) based on the 'forward' coral variable matrix (A) and the reconstruction target (b).
}
\examples{
# Load data from Hughes et al. (2024)
library(SMITER)
data(BMDA_1B_Comp)
data(BMDA_1B_EComp)

# Designate forward matrix (A), reconstruction target (b), and corresponding errors (Ae, be).

proxy_names <- names(bmda_1b_comp[,c(5:ncol(bmda_1b_comp))])

A <- bmda_1b_comp[,proxy_names]
Ae <- bmda_1b_ecomp[,proxy_names]
b <- bmda_1b_comp[,'Temp']
be <- rep(0.02, length(b))

# Execute SMITE Calibration with no truncation (i.e., all information retained).
result <- SMITE.calib(A = A, b = b, Ae = Ae, be = be, eigenclean = ncol(A))
}
